#9: I think I have it now, explanation below.
The problem here was that recent (4.x) versions of gcc tend to place certain parts of the code section into different sections than .text. This causes the output file to contain multiple code sections. This was not properly supported by the chaos ELF loader. I did a workaround for it, which didn't work, because the sections had to have a certain alignment (the workaround I did was simply too simplistic).

The proper way to solve this turned out to be absolutely trivial. :-)