More fixes for #14.
I think we still have problems with this. The "right" way would be to get the constraints right, but at the moment, I'm just too lazy and have too little time to feel it's worth it. Someone can always come along later, play with and submit a PR if they feel like improving on this. It's not like someone will run mission critical servers on chaos this decade anyway. :)