We want to be able to override certain methods, for testing purposes.
Weak symbols are quite useful for this. I looked into using the --wrap function in GNU ld instead, but the problem is that it only works for *undefined* functions, and I don't want to have to mess around with linking a special .a file just for the tests. So making them weak and just overriding them is simpler (since I have need for calling the underlying methods; at least not yet).