Added unfinished tests for memory_global_allocate and memory_global_deallocate.
This is unfortunately all but trivial. The problem is that the code hasn't really been written with testability in mind. Is it worth it? one may ask. "I don't know" is the simple answer. In some ways, yes: it makes sense to have an infrastructure for tests so we can easily (at a very low cost) add more tests later. But maybe this (memory_global_allocate et al) are just a few bad examples and it would make more sense to add tests for e.g. string_to_number and such - methods which have very few external dependencies etc.