Fixed the bug that caused the argument passing to fail. Explanation below.
The problem at hand here was, in the end, quite simple. We were trying to copy the stuff into the right location by creating a new pointer based on the ESP value minus the "stack base address". This was all in all correct, but the problem was that we were doing this while we had the *PL3* stack mapped. And that was the problem: we were using a memory address in the PL0 stack (since that's what ESP pointed to at the moment), so we were in essence writing to the wrong memory location...

Hard to find, but easy to fix once I managed to find it. Joy! :)