Restructured the FAT server entirely.
fat.c was incredibly long; around 1000 lines of code. I don't like that, so I decided to split it up into quite smallish files. More or less, one file per method (some methods are 200 lines long anyway, so it makes sense). To keep stuff easy to find, the names of the files match the method names in this case, except for fat.c which serves as entry point.

I also fixed the usage of system_thread_create() which is now different since the signature has changed.