Added NUMBER_OF_TIMESLICES kernelfs class.
This class lets us easily read the full number of timeslices (or rather, taskswitches) since the previous reset. The THREAD_INFO_VERBOSE call performs a reset, so in essence this gives us 'the number of timeslices/taskswitches since the user last read out all the thread info. In cluido, this is done using the 'top' command.