Updated to gcc version 5.
The update takes place not because we *depend* on anything that is new in gcc-5, but it's still good practice to upgrade early to find out any potential issues, and also because gcc 5 will sooner or later be the default (presumably). In this case, the upgrade was very smooth and trouble-free, which is nice.

----

Details:

* Vagrantfile: Updated to Debian 9 (a.k.a. stretch), so we can get the gcc-5 package (it't not available in jessie, i.e. Debian 8).
* rakelib/common_settings.rake: Use gcc-5 as the compiler by default.
* Rakefile: Don't build the iso_image task by default, since it will have garbage content in that case. It doesn't make sense to build it before the binaries have been installed in the right place.
* storm/ia32/dispatch.c: gcc-5 is more aggressive in its optimizations, so the  variable gets optimized away if we provide it as static.