Removed autochaos
We don't use this for anything any more. We have some libraries and servers which still rely on it for building, but they are not included in what we build at the moment. They will be converted to the Rake-based build pipeline shortly. Hence, let's drop this now once and forever, for the sake of cleaning up the campground a bit.