Removed the dead nibbles code
What do I mean with dead? Well, the `nibbles` command in Cluido hasn't been around for a while. And, the code had #ifdef:ed parts, parts that wouldn't compile if you would remove the #ifdef.

Simply put, this has become old cruft that is just now laying around gathering dust. It gives the _false_ impressions that it does something useful, which it doesnt. Hence, we must kill it with :fire:.

Let's not be sentimental with old code please. What I mean is: if someone wants to fix it, fine. I won't object. If so, we should:

- make it a proper "program" and put it under the `programs` folder, just like cluido and other programs.
- fix it so that it compiles.

We can even use this code as the basis for that. It's not like we are making it impossible or anything to fix it; we're just being very clear with the fact that _this is not working_ and should hence not be included in the `master` branch.