Vagrantfile: Default the ARCH to x86
raspberrypi is also available, but it's not supported for the whole system yet (and will probably not be for the next years). Let's default this to make the setup a bit more ergonomic. (People wishing to hack on the raspberrypi parts can just tweak their ~/.bashrc after `vagrant up && vagrant ssh`.)