Travis: Support newer gcc versions (#96)
* Travis: Support newer gcc versions

In fact, run on both gcc 4.9, 5, 6 and 7. This is probably overkill in the long run. If it turns out to be a maintenance nightmare, we can drop one or more of these easily.

* Attempt to fix warning w/ gcc 6 and 7.

* Attempt to fix gcc-7 compile error

Inspired by this: https://stackoverflow.com/questions/47981/how-do-you-set-clear-and-toggle-a-single-bit

* TIL: Explicit case fallthrough

More details: https://developers.redhat.com/blog/2017/03/10/wimplicit-fallthrough-in-gcc-7/

* Attempted workaround for -Wimplicit-fallthrough

The comment approach did not work for me; the warning level seemed not to be 3 as it ought to be. Maybe if we explicitly set it to 4, things will work.

* ARCH_CFLAGS was only being used when compiling chaos.

Should be used when compiling the libraries and servers also.

* Try level 3 instead.

* One more try to get it working.

* Googled for -Wimplicit-fallthrough

This should work. The commment must be outside of the block, if it exists...

* One more futile attempt to get it working w/ gcc-7

* One more try

Inspired by https://github.com/ponylang/ponyc/issues/1581