Linker scripts: made it possible to override PROCESS_VM_BASE
This turned out to be crucial to be able to debug the issues I was having tonight, so I put the time into fixing this. It turned out to be quite trivial in the end, since a Stack Overflow user had done the vast majority of the work for us. Thanks Aravind!

Closes #104.