Bug fix: The FAT server should handle modern FAT16 volumes (#111)
The number of reserved sectors _used_ to be always 1, but this is no longer the case. The FAT volume we format ourselves with `mkdosfs` no longer follows this age-old convention, which is mentioned in e.g. the classical FAT: General Overview of On-Disk Format specification by Microsoft. So we must read this value from the BPB instead of just blindly making foolish assumptions.