[servers/fat] Fix support for case flag (#130)
This fixes #107. It was actually a bit more work to fix than just checking if the flag was 0 or 1, but still, rather easy which can be seen at the length of this diff.

This page helped me a lot understanding the format: http://elm-chan.org/docs/fat_e.html. GDB was also helpful, as usual.