[servers/pci] Disable PCI setup for 0x8086/0x9d23
This is a known "bad device" in terms of our PCI setup code. Or rather; our PCI setup code breaks with this device, likely because it hasn''t caught up with the last 20 years of development in the PC world. :-)

This will do for now; I have verified on the machine in question that we don't reboot on startup when this device is exempt from the PCI setup.

Issue for fixing this long-term: #134.