[build] Add support for building Docker images (#136)
This is pretty "quick and dirty", but works reasonably well. The code is
both built and run inside two separate Docker containers - one with gcc,
nasm etc (with an obviously larger footprint) and another one based on a
smaller Debian base image, with only qemu and the resulting `.iso` file.