[servers/console] Add new activation flag (#137)
This flag is to be used by cluido, and by cluido only, to ensure that
the cluido console is always made visible when created; other servers
have a lower priority. This is important to give the user a decent
experience when running e.g. the chaos Docker container, using a
pre-compiled chaos .iso image etc.

Closes #45; there are still some issues with the kernel log console (as
being created by the log server) not being visible, but I will create a
separate issue for that before merging this one.