[servers/fat] Fix fat_file_open case bug
I discovered another incarnation of #107. The previous fix (in #130)
fixed `fat_file_get_info`, _but_ the code was still broken for
`fat_file_open`. There was also a weird bug making "no matching file"
use the last file that happened to exist in the directory suggested (!).

In addition to all of this, `fat_file_open` stupidly duplicated code
already present elsewhere. All of this is fixed in this commit.

This was the root cause for #132. In other words: Fixes #132.