[storm] Refactor IPC to use circular queue
This was a rather complex change which was discussed with @doverhill on
Slack a bit. Because our memory allocation tend to degenerate over time,
we looked into other ways of handling the IPC than the previous approach
(which would incur a memory allocation and deallocation for _every
single message_ being posted via the IPC mechanism)

The new approach uses a circular queue, which was written in an isolated
and reusable way, completely separate from the mailbox code. This means
that `mailbox_send` and `mailbox_receive` got a whole lot more readable
in the process, as an added bonus.

The implementation was greatly helped by the fact that we could unit
test the circular queue in isolation, in VS Code on Linux (with full
`gdb` debugger support). This was an _incredible_ help in getting this
done with a reasonable level of effort.

Kernel development has never felt this "easy" before, relatively
speaking. I still ran into a bug (in `mailbox_receive`) but it was
literally a 10-minute fix. After it was done, the system would boot up
correctly again, which felt really nice.