[programs/modplay] Add .mod visualization
This was a bit tricky, but once I figured out that just rendering the
PCM value (shortening it to the appropriate number of bits) would look
reasonable, it wasn't so hard. The tricky part this time was finding a
good audio visualization algorithm.

Also added a listing of the modfiles available, and made it possible to
pick a file by using the arrow keys. This is interestingly enough a bit
laggy, I did some attempts to reduce the lag but... didn't really
succeed. I _did_ suceed in getting the audio playback to stutter,
though, and I really don't want that so... this will have to be good
enough for now.