[docker] Unbreak build by forcing use of gcc-7
The fix in b5f6d16 took us part of the way, but since `gcc-8` is default
in Debian Buster it was inconclusive; the build would still be attempted
with `gcc-8`, which fails until #118 gets merged.

In the meanwhile, this fixes Docker building to use `gcc-7` instead
(which is the version I still use locally btw)

In the longer run, we could also consider doing a `docker build` in
Travis. Yes, it's one extra job, but the problem with the Docker Hub job
is that I currently get no notification if/when its builds are failing.
In this case, they had been failing for weeks/months without anyone
noticing, which is a bit bad.