(pci) Include vendor and device ID in pci_device_info_type
The use case is this: For device drivers which supports multiple PCI
device ID:s, it is sometimes important to be able to differentiate minor
details in the hardware-related code depending on device ID. For
example, the EEPROM of a network card might be read using slightly
different mechanisms on different device ID:s, whereas the rest of the
hardware support (PHY etc) is entirely identical.

So, to be able to better support such use cases, we are now adding a few
fields to the pci_device_info_type, which is used when probing for PCI
devices.