(storm) Fix debug_print() handling of large unsigned integers
The problem was that the `decimal_string()` helper method being used
would take a signed integer as a parameter, which made it impossible to
print integers with bit 31 set (since this is the sign bit in
two's complement representation of signed numbers).

Being able to use unit tests for this made the debugging experience
short and sweet, just like it ought to be. Adding VSCode configs for
this, since I noted they were not present in the repo.